requirements.txt
torch
torchvision
pillow
huggingface_hub